feat: change upload lambda to use presigned url, address PR comments
