--- lambda/aws-rag-appsync-stepfn-kendra/input_validation/src/requirements.txt
@@ -1,4 +1,0 @@
-aws-lambda-powertools
-boto3
-requests
-requests-aws4auth
