--- lambda/aws-rag-appsync-stepfn-kendra/s3_file_uploader/src/requirements.txt
@@ -1,4 +1,0 @@
-aws-lambda-powertools
-typing-extensions
-boto3
-requests
