+++ src/common/helpers/lambda-builder-helper.ts
